requirements.txt
# Core dependencies
pypinyin>=0.50.0

# Vector search and RAG (optional)
numpy>=1.24.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Document parsing (optional)
PyPDF2>=3.0.0

# OpenAI API (optional, for LLM-powered Q&A)
openai>=1.0.0
